requirements.txt
apipkg==1.5
astroid==2.2.0
atomicwrites==1.3.0
attrs==18.2.0
beautifulsoup4==4.7.1
brazilnum==0.8.8
certifi==2018.11.29
chardet==3.0.4
coverage==4.5.3
Django==2.2.4
django-auth-ldap==1.7.0
django-bootstrap4==0.0.7
django-cas-ng==3.6.0
django-ckeditor==5.7.1
django-easy-audit==1.1
django-extensions==2.1.9
django-filter==2.2.0
django-js-asset==1.2.2
django-nose==1.4.6
django-recaptcha==2.0.4
execnet==1.6.0
Faker==0.9.1
idna==2.8
importlib-metadata==0.18
isort==4.3.9
lazy-object-proxy==1.3.1
ldap3==2.5.2
lxml==4.3.1
mccabe==0.6.1
mixer==6.1.3
more-itertools==6.0.0
nose==1.3.7
packaging==19.0
pep8==1.7.1
Pillow==5.4.1
pluggy==0.12.0
psycopg2==2.7.7
psycopg2-binary==2.7.7
py==1.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pydot==1.4.1
pygraphviz==1.5
pylint==2.3.0
pylint-django==2.0.5
pylint-plugin-utils==0.4
pyparsing==2.4.0
pytest==4.6.3
pytest-cache==1.0
pytest-django==3.4.8
pytest-forked==1.0.2
pytest-pep8==1.0.6
pytest-xdist==1.29.0
python-cas==1.4.0
python-dateutil==2.8.0
python-ldap==3.1.0
python3-ldap==0.9.8.4
pytz==2018.9
requests==2.21.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
text-unidecode==1.2
typed-ast==1.3.1
urllib3==1.24.1
validate-email==1.3
wcwidth==0.1.7
wrapt==1.11.1
zipp==0.5.1
